Time: 56 ms (48.03%), Space: 30.1 MB (27.13%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -16,6 +16,6 @@ public:
                 else dp[i][j] = dp[i-1][j];
             }
         }
-        return (int)dp[n][m];
+        return dp[n][m];
     }
 };
